Change Wait decorator configuration internals
Because decorator instances are created when the code of decorated
class/method is read, code taking care of configuration doesn't
have a chance to run before that. Due to that fact, configuration
responsibility has been of loaded to private class and Wait decorator
now behavies as semi-singleton. Configuration and logger is contained
in private class "static instance" and decorated method is sontained
in decorator instance it self.

It allows to configure decorator globally even thou we don't have
references for individual decorator instances.